requirements.txt
streamlit
plotly
python-dotenv
groq
python-magic-bin